+++ LazyPizzaApp/src/main/java/pl/lazypizza/domain/model/Product.kt
@@ -9,7 +9,7 @@ data class Product(
     val id:String = "",
     val name: String = "",
     val description: String = "",
-    val price: Double = 0.0,
+    val price: Float = 0f,
 
     @get:PropertyName("category")
     @set:PropertyName("category")
